APT 21: Step 7: Show a Random Restaurant
Added a WidgetService class that rotates through random restaurants in the widget, also edited a few string resources, one for if there is an empty set of restuarants.
Also added the WidgetService to the manifest and fixed a few problems with my detail_form layout and how the notes and feed fields were resolving the layout.